fix: disable pthread_atfork fork detection on OpenBSD
The pthread_atfork fork detection mechanism uncovered a bug in OpenBSD's
libc which is triggered when a process creates multiple generations of
children, and is multi-threaded, and calls pthread_atfork(). When the
grandchild process tries to fork, it deadlocks on the atfork lock. More
information can be found here
https://marc.info/?l=openbsd-tech&m=167047636422884&w=2

While a patch has been sent to OpenBSD which remediates the deadlock, it
so far hasn't received much attention and won't be committed any time
soon. Per #3701, it's deemed an acceptable risk to disable the atfork
mechanism in s2n-tls since there are still other layers of defense in
place.